+++ tests/benchmarks/dataclasses/test_benchmark_saga_memory.py
@@ -179,7 +179,7 @@ def saga_with_memory_storage() -> Saga[OrderContext]:
 
 
 @pytest.mark.benchmark
-def test_benchmark_saga_memory_full_transaction(
+def test_benchmark_saga_memory_run_full_transaction(
     benchmark,
     saga_with_memory_storage: Saga[OrderContext],
     saga_container: SagaContainer,
@@ -206,7 +206,7 @@ def test_benchmark_saga_memory_full_transaction(
 
 
 @pytest.mark.benchmark
-def test_benchmark_saga_memory_single_step(
+def test_benchmark_saga_memory_run_single_step(
     benchmark,
     saga_with_memory_storage: Saga[OrderContext],
     saga_container: SagaContainer,
