+++ tests/integration/test_saga_mediator_sqlalchemy_postgres.py
@@ -1,28 +1,19 @@
 """Integration tests for SagaMediator with SqlAlchemySagaStorage (PostgreSQL)."""
 
-import typing
 import uuid
-from unittest import mock
 
 import pytest
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 import cqrs
-from cqrs import events
-from cqrs.requests.map import SagaMap
 from cqrs.saga.storage.enums import SagaStatus
 from cqrs.saga.storage.sqlalchemy import SqlAlchemySagaStorage
 
 from tests.integration.test_saga_mediator_memory import (
     FailingOrderSaga,
-    FailingStep,
-    InventoryReservedEvent,
     InventoryReservedEventHandler,
     OrderContext,
-    OrderShippedEvent,
     OrderShippedEventHandler,
-    OrderSaga,
-    PaymentProcessedEvent,
     PaymentProcessedEventHandler,
     ProcessPaymentResponse,
     ProcessPaymentStep,
@@ -33,104 +24,12 @@ from tests.integration.test_saga_mediator_memory import (
 )
 
 
-class _TestContainer:
-    """Test container that resolves step handlers, sagas, and event handlers."""
-
-    def __init__(self, storage: SqlAlchemySagaStorage) -> None:
-        self._storage = storage
-        self._external_container = None
-        self._step_handlers = {
-            ReserveInventoryStep: ReserveInventoryStep(),
-            ProcessPaymentStep: ProcessPaymentStep(),
-            ShipOrderStep: ShipOrderStep(),
-            FailingStep: FailingStep(),
-        }
-        self._event_handlers = {
-            InventoryReservedEventHandler: InventoryReservedEventHandler(),
-            PaymentProcessedEventHandler: PaymentProcessedEventHandler(),
-            OrderShippedEventHandler: OrderShippedEventHandler(),
-        }
-        self._sagas = {
-            OrderSaga: OrderSaga(),  # type: ignore[arg-type]
-            FailingOrderSaga: FailingOrderSaga(),  # type: ignore[arg-type]
-        }
-
-    @property
-    def external_container(self) -> typing.Any:
-        return self._external_container
-
-    def attach_external_container(self, container: typing.Any) -> None:
-        self._external_container = container
-
-    async def resolve(self, type_) -> typing.Any:
-        if type_ in self._step_handlers:
-            return self._step_handlers[type_]
-        if type_ in self._event_handlers:
-            return self._event_handlers[type_]
-        if type_ in self._sagas:
-            return self._sagas[type_]
-        if type_ == SqlAlchemySagaStorage:
-            return self._storage
-        raise ValueError(f"Unknown type: {type_}")
-
-
 @pytest.fixture
 def storage(
     saga_session_factory_postgres: async_sessionmaker[AsyncSession],
 ) -> SqlAlchemySagaStorage:
     """Create SqlAlchemySagaStorage instance (PostgreSQL)."""
     return SqlAlchemySagaStorage(saga_session_factory_postgres)
-
-
-@pytest.fixture
-def container(storage: SqlAlchemySagaStorage) -> _TestContainer:
-    """Create test container."""
-    container = _TestContainer(storage)
-    for step_handler in container._step_handlers.values():
-        if hasattr(step_handler, "_events"):
-            step_handler._events.clear()
-    for event_handler in container._event_handlers.values():
-        if hasattr(event_handler, "handled_events"):
-            event_handler.handled_events.clear()
-    return container
-
-
-@pytest.fixture
-def saga_mediator(
-    container: _TestContainer,
-    storage: SqlAlchemySagaStorage,
-) -> cqrs.SagaMediator:
-    """Create SagaMediator with SqlAlchemySagaStorage (PostgreSQL)."""
-
-    def saga_mapper(mapper: SagaMap) -> None:
-        mapper.bind(OrderContext, OrderSaga)
-
-    def events_mapper(mapper: events.EventMap) -> None:
-        mapper.bind(InventoryReservedEvent, InventoryReservedEventHandler)
-        mapper.bind(PaymentProcessedEvent, PaymentProcessedEventHandler)
-        mapper.bind(OrderShippedEvent, OrderShippedEventHandler)
-
-    event_map = events.EventMap()
-    events_mapper(event_map)
-    message_broker = mock.AsyncMock()
-    message_broker.produce = mock.AsyncMock()
-    event_emitter = events.EventEmitter(
-        event_map=event_map,
-        container=container,  # type: ignore
-        message_broker=message_broker,
-    )
-    saga_map = SagaMap()
-    saga_mapper(saga_map)
-    mediator = cqrs.SagaMediator(
-        saga_map=saga_map,
-        container=container,  # type: ignore
-        event_emitter=event_emitter,
-        event_map=event_map,
-        max_concurrent_event_handlers=2,
-        concurrent_event_handle_enable=True,
-        storage=storage,
-    )
-    return mediator
 
 
 class TestSagaMediatorSqlAlchemyStoragePostgres:
